requirements.txt
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
awscli==1.15.37
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
boto3==1.7.33
botocore==1.10.37
Bottleneck==1.2.1
bs4==0.0.1
carmenproxy==0.1.0
certifi==2018.8.24
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
conda==4.4.10
conda-build==3.4.1
conda-verify==2.0.0
contextlib2==0.5.5
coverage==4.5.1
coveralls==1.3.0
cryptography==2.1.4
custom-inherit==2.2.0
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
d3m==2018.7.10
dask==0.16.1
datashape==0.5.4
dateparser==0.7.0
decorator==4.2.1
distributed==1.20.2
Django==1.11.13
docopt==0.6.2
docutils==0.14
easydict==1.7
elasticsearch==6.3.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.10
fastcache==1.0.2
filelock==2.0.13
Flask==0.12.2
Flask-Cors==3.0.3
frozendict==1.2
future==0.16.0
gevent==1.2.2
gitdb2==2.0.4
GitPython==2.1.9
glob2==0.6
gmpy2==2.0.8
graphviz==0.8.3
greenlet==0.4.12
gym==0.10.8
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.8.0
ipytest==0.3.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonpath-ng==1.4.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
jupyterlab==0.31.5
jupyterlab-launcher==0.10.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
ldap3==2.5
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
mpmath==1.0.0
msgpack-python==0.5.1
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
nose-timer==0.7.2
notebook==5.4.0
ntlm-auth==1.2.0
numba==0.36.2
numexpr==2.6.4
numpy==1.14.3
numpydoc==0.7.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.4.10
packaging==16.8
pandas==0.23.4
pandas-summary==0.0.41
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pbr==4.0.4
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
pkginfo==1.4.2
placebo==0.8.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pydocstyle==3.0.0
pyflakes==1.6.0
pyglet==1.2.4
Pygments==2.2.0
pylama==7.6.5
pylama-pylint==3.1.1
pylint==1.8.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
python-ldap==3.1.0
python-magic==0.4.15
python-twitter==3.4.2
pytypes==1.0b4
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
regex==2018.7.11
requests==2.18.4
requests-ntlm==1.1.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rfc3987==1.3.7
rope==0.10.7
rsa==3.4.2
ruamel-yaml==0.15.35
s3cmd==2.0.2
s3fs==0.1.5
s3transfer==0.1.13
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.9.0
selenium==3.14.0
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smmap2==2.0.4
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.2.6
SQLAlchemy==1.2.1
statsmodels==0.9.0
strict-rfc3339==0.7
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
tika==1.18
toolz==0.9.0
torch==0.4.1
torch-vision==0.1.6.dev0
torchvision==0.2.1
tornado==4.5.3
tox==3.0.0
tqdm==4.23.1
traitlets==4.3.2
twine==1.11.0
twitter==1.18.0
typing==3.6.2
typing-inspect==0.2.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
webcolors==1.8.1
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wordcloud==1.5.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.2.0
zict==0.1.3
